requirements.txt
boto3==1.7.6
botocore==1.10.6
docutils==0.14
jmespath==0.9.3
python-dateutil==2.7.2
PyYAML==3.12
s3transfer==0.1.13
six==1.11.0
